main

+++ cell02/ex00/iszero.py
@@ -1,4 +1,4 @@
-n = int(input())
+n = float(input())
 
 if n == 0:
     print("This number is equal to zero.")
